refactor: Сделать модульность
Сделано модульное разделение
