+++ js/main.js
@@ -1,41 +1,2 @@
-const getRandomInt = (min, max) =>
-  Math.floor(Math.random() * (max - min + 1)) + min;
-
-const messages = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
-
-const names = [
-  'Артём',
-  'Мария',
-  'Иван',
-  'Ольга',
-  'Дмитрий',
-  'Светлана',
-];
-
-const generateComment = (id) => ({
-  id,
-  avatar: `./img/avatar-${getRandomInt(1, 6)}.svg`,
-  message: messages[getRandomInt(0, messages.length - 1)],
-  name: names[getRandomInt(0, names.length - 1)],
-});
-
-const generateComments = () =>
-  Array.from({ length: getRandomInt(1, 30) }, (_, i) => generateComment(i + 1));
-
-const generatePhotos = () =>
-  Array.from({ length: 25 }, (_, i) => ({
-    id: i + 1,
-    url: `./photos/${i + 1}.jpg`,
-    description: `Описание фотографии #${i + 1}`,
-    likes: getRandomInt(15, 200),
-    comments: generateComments(),
-  }));
-
+import { generatePhotos } from './util.js';
 generatePhotos();
